requirements.txt
imageio
imagesize
numpy
openpyxl
pandas
Pillow==8.4.0
plotly==5.6.0
streamlit==1.5.1
XlsxWriter